refactor: Slogan폴더내부로 컴포넌트 파일 변경 #653
