+++ frontend/src/pages/landing/Landing.tsx
@@ -4,7 +4,7 @@ import landing1 from '../../common/assets/images/landing1.png';
 import landing2 from '../../common/assets/images/landing2.png';
 import landing3 from '../../common/assets/images/landing3.png';
 
-import Slogan from './components/Slogan';
+import Slogan from './components/Slogan/Slogan';
 
 function Landing() {
   return (
